// packages/orchestration/src/client/api/schema/tool-call-chunk.ts
/*
 * Copyright (c) 2026 SAP SE or an SAP affiliate company. All rights reserved.
 *
 * This is a generated file powered by the SAP Cloud SDK for JavaScript.
 */

/**
 * Representation of the 'ToolCallChunk' schema.
 */
export type ToolCallChunk = {
  index: number;
  /**
   * The ID of the tool call.
   */
  id?: string;
  /**
   * The type of the tool. Currently, only `function` is supported.
   */
  type?: 'function';
  function?: {
    /**
     * The name of the function to call.
     */
    name?: string;
    /**
     * The arguments to call the function with, as generated by the model in JSON format. Note that the model does not always generate valid JSON, and may hallucinate parameters not defined by your function schema. Validate the arguments in your code before calling your function.
     */
    arguments?: string;
  } & Record<string, any>;
} & Record<string, any>;